Update to css-draft
I styled h1(#title) and h2(#text)- the title and quote. I added some padding to get #title in the perfect position, so it was just in a similar position to the actual JLR page. Both #text and #title have a lighter font-weight than the one used in the navigation bar. Will continue adding to this!